Extract ingestion unit of work abstraction
